server/requirements.txt
alembic==1.15.2
aniso8601==10.0.1
blinker==1.9.0
CacheControl==0.14.3
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
cryptography==44.0.3
distlib==0.3.9
filelock==3.18.0
firebase-admin==6.8.0
Flask==3.1.0
flask-cors==5.0.1
Flask-JWT-Extended==4.7.1
Flask-Migrate==4.1.0
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
google-api-core==2.24.2
google-api-python-client==2.169.0
google-auth==2.39.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.3
google-cloud-firestore==2.20.2
google-cloud-storage==3.1.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
greenlet==3.2.1
grpcio==1.71.0
grpcio-status==1.71.0
httplib2==0.22.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.2
msgpack==1.1.0
packaging==25.0
pipenv==2025.0.1
platformdirs==4.3.7
proto-plus==1.26.1
protobuf==5.29.4
psycopg2-binary==2.9.9
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
PyJWT==2.10.1
pyparsing==3.2.3
python-dotenv==1.1.0
pytz==2024.2
requests==2.32.3
rsa==4.9.1
six==1.17.0
SQLAlchemy==2.0.29
sqlalchemy-serializer==1.4.22
typing_extensions==4.13.2
uritemplate==4.1.1
urllib3==2.4.0
virtualenv==20.30.0
Werkzeug==3.1.3
gunicorn
